update utils and constants
