+++ packages/providers/google/package.json
@@ -38,6 +38,6 @@
     "@llamaindex/env": "workspace:*"
   },
   "dependencies": {
-    "@google/genai": "^1.6.0"
+    "@google/genai": "^1.7.0"
   }
 }
