+++ packages/providers/google/src/base.ts
@@ -18,6 +18,12 @@ import type {
 import { ToolCallLLM } from "@llamaindex/core/llms";
 import { streamConverter } from "@llamaindex/core/utils";
 import { getEnv, randomUUID } from "@llamaindex/env";
+import {
+  DEFAULT_GEMINI_PARAMS,
+  DEFAULT_SAFETY_SETTINGS,
+  GEMINI_MODEL_INFO_MAP,
+  SUPPORT_TOOL_CALL_MODELS,
+} from "./constants.js";
 import { GeminiLive } from "./live.js";
 import {
   GEMINI_MODEL,
@@ -27,65 +33,9 @@ import {
   type GeminiChatParamsStreaming,
   type GeminiChatStreamResponse,
   type GeminiMessageRole,
-  type GeminiModelInfo,
   type GeminiVoiceName,
 } from "./types.js";
-import {
-  DEFAULT_SAFETY_SETTINGS,
-  GeminiHelper,
-  getChatContext,
-  getPartsText,
-} from "./utils.js";
-
-export const GEMINI_MODEL_INFO_MAP: Record<GEMINI_MODEL, GeminiModelInfo> = {
-  [GEMINI_MODEL.GEMINI_PRO]: { contextWindow: 30720 },
-  [GEMINI_MODEL.GEMINI_PRO_VISION]: { contextWindow: 12288 },
-  // multi-modal/multi turn
-  [GEMINI_MODEL.GEMINI_PRO_LATEST]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_PRO_FLASH_LATEST]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_PRO_1_5_PRO_PREVIEW]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_PRO_1_5_FLASH_PREVIEW]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_PRO_1_5]: { contextWindow: 2 * 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_PRO_1_5_FLASH]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_PRO_1_5_LATEST]: { contextWindow: 2 * 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_PRO_1_5_FLASH_LATEST]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_0_FLASH_EXPERIMENTAL]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_0_FLASH]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_0_FLASH_LITE_PREVIEW]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_0_FLASH_LITE]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_0_FLASH_LIVE]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_0_FLASH_THINKING_EXP]: { contextWindow: 32768 },
-  [GEMINI_MODEL.GEMINI_2_0_PRO_EXPERIMENTAL]: { contextWindow: 2 * 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_5_PRO_PREVIEW]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_5_PRO_PREVIEW_LATEST]: { contextWindow: 10 ** 6 },
-  [GEMINI_MODEL.GEMINI_2_5_FLASH_PREVIEW]: { contextWindow: 10 ** 6 },
-};
-
-export const SUPPORT_TOOL_CALL_MODELS: GEMINI_MODEL[] = [
-  GEMINI_MODEL.GEMINI_PRO,
-  GEMINI_MODEL.GEMINI_PRO_VISION,
-  GEMINI_MODEL.GEMINI_PRO_1_5_PRO_PREVIEW,
-  GEMINI_MODEL.GEMINI_PRO_1_5_FLASH_PREVIEW,
-  GEMINI_MODEL.GEMINI_PRO_1_5,
-  GEMINI_MODEL.GEMINI_PRO_1_5_FLASH,
-  GEMINI_MODEL.GEMINI_PRO_LATEST,
-  GEMINI_MODEL.GEMINI_PRO_FLASH_LATEST,
-  GEMINI_MODEL.GEMINI_PRO_1_5_LATEST,
-  GEMINI_MODEL.GEMINI_PRO_1_5_FLASH_LATEST,
-  GEMINI_MODEL.GEMINI_2_0_FLASH_EXPERIMENTAL,
-  GEMINI_MODEL.GEMINI_2_0_FLASH,
-  GEMINI_MODEL.GEMINI_2_0_PRO_EXPERIMENTAL,
-  GEMINI_MODEL.GEMINI_2_5_PRO_PREVIEW,
-  GEMINI_MODEL.GEMINI_2_5_PRO_PREVIEW_LATEST,
-  GEMINI_MODEL.GEMINI_2_5_FLASH_PREVIEW,
-];
-
-export const DEFAULT_GEMINI_PARAMS = {
-  model: GEMINI_MODEL.GEMINI_PRO,
-  temperature: 0.1,
-  topP: 1,
-  maxTokens: undefined,
-};
+import { GeminiHelper, getChatContext, getPartsText } from "./utils.js";
 
 export type GeminiConfig = Partial<typeof DEFAULT_GEMINI_PARAMS> & {
   safetySettings?: SafetySetting[];
